+++ NEWS.md
@@ -1,3 +1,5 @@
+# openmpp (development version)
+
 # openmpp 0.0.1
 
 * Initial public release.