+++ NEWS.md
@@ -1,4 +1,4 @@
-# openmpp (development version)
+# openmpp 0.0.2
 
 * We have added new functionality for users to specify their own custom connections to the OpenM++ OMS API. This should make the package more flexible for different types of user authentication.